Lesson -10- Refactor code
