+++ src/test/java/parameterizeTests/helper/TestDataWithFaker.java
@@ -1,4 +1,4 @@
-package parameterizeTests;
+package parameterizeTests.helper;
 
 import com.github.javafaker.Faker;
